+++ .github/workflows/go-test.yaml
@@ -30,7 +30,7 @@ jobs:
           skip-build-cache: true
 
       - name: Run tests
-        run: TEST_PARAM="-coverprofile=coverage.out -covermode=atomic" make test
+        run: TEST_PARAM="-coverprofile=coverage.out -covermode=atomic" make test-all
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
