+++ driver/search.go
@@ -16,6 +16,8 @@ package driver
 
 import (
 	"context"
+	"errors"
+	"io"
 
 	api "github.com/tigrisdata/tigris-client-go/api/server/v1"
 )
@@ -42,4 +44,59 @@ type SearchClient interface {
 	Delete(ctx context.Context, name string, ids []string) ([]*DocStatus, error)
 	DeleteByQuery(ctx context.Context, name string, filter Filter) (int32, error)
 	Search(ctx context.Context, name string, req *SearchRequest) (SearchIndexResultIterator, error)
+}
+
+type SearchIndexResultIterator interface {
+	Next(r *SearchIndexResponse) bool
+	Err() error
+	Close()
+}
+
+type searchIndexReader interface {
+	read() (SearchIndexResponse, error)
+	close() error
+}
+
+type searchIndexResultIterator struct {
+	searchIndexReader
+	eof bool
+	err error
+}
+
+func (i *searchIndexResultIterator) Next(r *SearchIndexResponse) bool {
+	if i.eof {
+		return false
+	}
+
+	resp, err := i.read()
+	if errors.Is(err, io.EOF) {
+		i.eof = true
+		_ = i.close()
+		return false
+	}
+
+	if err != nil {
+		i.eof = true
+		i.err = err
+		_ = i.close()
+
+		return false
+	}
+
+	*r = resp
+
+	return true
+}
+
+func (i *searchIndexResultIterator) Err() error {
+	return i.err
+}
+
+func (i *searchIndexResultIterator) Close() {
+	if i.eof {
+		return
+	}
+
+	_ = i.close()
+	i.eof = true
 }
