+++ driver/search_http.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build tigris_http
+
 //nolint:bodyclose
 package driver
