+++ search/index_test.go
@@ -445,7 +445,7 @@ func TestSearchOpen(t *testing.T) {
 	ctx, cancel1 := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel1()
 
-	cfg := &Config{URL: test.GRPCURL(8), Project: "db1"}
+	cfg := &Config{URL: test.URL(8), Project: "db1"}
 	cfg.TLS = test.SetupTLS(t)
 
 	ms.Search.EXPECT().CreateOrUpdateIndex(gomock.Any(),
